Split cell steps
